requirements.txt
Django>=4.1.7,<4.2.0
djangorestframework>=3.13.0,<3.14.0
flake8>=6.0.0,<6.1.0
psycopg2-binary>=2.9.5,<2.10.0
drf-spectacular>=0.16,<0.17
djangorestframework-simplejwt>=5.3.0,<5.4.0
Pillow>=10.1.0,<10.2.0
